requirements.txt
alembic==1.8.1
anyio==3.6.1
appdirs==1.4.4
click==8.0.3
DateTime==4.3
distlib==0.3.2
ez-setup==0.9
fastapi==0.85.0
FastAPI-SQLAlchemy==0.2.1
filelock==3.0.12
h11==0.13.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
Mako==1.2.2
MarkupSafe==2.1.1
nltk==3.6.7
numpy==1.22.0
pandas==1.4.0
pika==1.3.0
pkgconfig==1.5.5
psycopg2==2.9.3
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2021.1
regex==2021.11.10
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.40
starlette==0.20.4
tqdm==4.62.3
typing_extensions==4.3.0
uvicorn==0.18.3
virtualenv==20.4.7
Werkzeug==2.2.2
zipp==3.8.1
zope.interface==5.4.0
